backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.5.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator==2.1.1
python-dotenv==1.0.1
certifi>=2024.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
cryptography>=42.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
fuzzywuzzy==0.18.0
python-Levenshtein>=0.25.0
python-dateutil>=2.9.0
httpx>=0.27.0
Pillow>=11.0.0
pymongo>=4.0.0
